getFormJSONConfig
